reorganize some of the files
