+++ packages/teraslice-test-harness/test/fixtures/asset/simple-connector/client.ts
@@ -15,4 +15,8 @@ export = class SimpleClient {
             count
         };
     }
+
+    isFinished() {
+        return false;
+    }
 };
